chore(svm): remove `send_transaction` action to codec
