+++ addons/svm/src/codec/mod.rs
@@ -1,6 +1,7 @@
 pub mod anchor;
 pub mod idl;
 pub mod instruction;
+pub mod send_transaction;
 
 use bip39::Language;
 use bip39::Mnemonic;
