+++ addons/svm/src/commands/mod.rs
@@ -4,7 +4,6 @@ use deploy_program::DEPLOY_PROGRAM;
 use process_instructions::PROCESS_INSTRUCTIONS;
 use send_sol::SEND_SOL;
 use send_token::SEND_TOKEN;
-use send_transaction::SEND_TRANSACTION;
 use sign_transaction::SIGN_TRANSACTION;
 use txtx_addon_kit::types::commands::PreCommandSpecification;
 use txtx_addon_kit::types::stores::ValueStore;
@@ -14,7 +13,6 @@ pub mod deploy_program;
 pub mod process_instructions;
 pub mod send_sol;
 pub mod send_token;
-pub mod send_transaction;
 pub mod sign_transaction;
 
 fn get_signers_did(args: &ValueStore) -> Result<Vec<ConstructDid>, Diagnostic> {
@@ -34,7 +32,6 @@ lazy_static! {
     pub static ref ACTIONS: Vec<PreCommandSpecification> = vec![
         SIGN_TRANSACTION.clone(),
         // ENCODE_INSTRUCTION.clone(),
-        SEND_TRANSACTION.clone(),
         PROCESS_INSTRUCTIONS.clone(),
         DEPLOY_PROGRAM.clone(),
         SEND_SOL.clone(),
