+++ addons/svm/src/signers/secret_key.rs
@@ -28,7 +28,7 @@ use txtx_addon_kit::types::{
     types::{Type, Value},
 };
 
-use crate::commands::send_transaction::send_transaction;
+use crate::codec::send_transaction::send_transaction;
 use crate::constants::{
     ACTION_ITEM_CHECK_ADDRESS, ACTION_ITEM_PROVIDE_SIGNED_TRANSACTION, ADDRESS, CHECKED_ADDRESS,
     CHECKED_PUBLIC_KEY, COMMITMENT_LEVEL, DO_AWAIT_CONFIRMATION, IS_DEPLOYMENT, IS_SIGNABLE,
